+++ docs/_static/.gitkeep
@@ -1,0 +1,2 @@
+# Static files directory for Sphinx
+# This directory should contain static files like CSS, JS, images, etc.
